+++ lib/src/models/models.dart
@@ -1,1 +1,2 @@
 export 'yaml_comment.dart';
+export 'yaml_break_line.dart';